+++ custom_components/linknlink/switch.py
@@ -16,9 +16,12 @@ from aiolinknlink import (
     PID_DTU,
     PID_EAC1_PANEL,
     PID_SINGLE_CHANNEL_LIGHT_SWITCH,
+    PID_THREE_CHANNEL_LIGHT_SWITCH,
     PID_TWO_CHANNEL_LIGHT_SWITCH,
     SINGLE_CHANNEL_LIGHT_BACKLIGHT_FIELD,
     SINGLE_CHANNEL_LIGHT_POWER_FIELD,
+    THREE_CHANNEL_LIGHT_BACKLIGHT_FIELD,
+    THREE_CHANNEL_LIGHT_MASTER_POWER_FIELD,
     TWO_CHANNEL_LIGHT_BACKLIGHT_FIELD,
     TWO_CHANNEL_LIGHT_MASTER_POWER_FIELD,
 )
@@ -69,6 +72,20 @@ TWO_CHANNEL_LIGHT_SWITCHES = (
         translation_key="all_switches",
     ),
 )
+THREE_CHANNEL_LIGHT_SWITCHES = (
+    *BOX7_SWITCHES[:3],
+    SwitchEntityDescription(
+        key=THREE_CHANNEL_LIGHT_BACKLIGHT_FIELD,
+        name="Panel backlight",
+        translation_key="panel_backlight",
+        icon="mdi:lightbulb-outline",
+    ),
+    SwitchEntityDescription(
+        key=THREE_CHANNEL_LIGHT_MASTER_POWER_FIELD,
+        name="All switches",
+        translation_key="all_switches",
+    ),
+)
 EAC1_SWITCHES = (
     SwitchEntityDescription(
         key=EAC1_KEY_LOCK_FIELD,
@@ -84,6 +101,7 @@ SWITCHES_BY_PID = {
     PID_EAC1_PANEL: EAC1_SWITCHES,
     PID_SINGLE_CHANNEL_LIGHT_SWITCH: SINGLE_CHANNEL_LIGHT_SWITCHES,
     PID_TWO_CHANNEL_LIGHT_SWITCH: TWO_CHANNEL_LIGHT_SWITCHES,
+    PID_THREE_CHANNEL_LIGHT_SWITCH: THREE_CHANNEL_LIGHT_SWITCHES,
 }